create tasklist class
