+++ src/main/java/duke/Deadline.java
@@ -1,7 +1,9 @@
+package duke;
+
 import java.time.LocalDate;
 
 public class Deadline extends Task {
-    protected LocalDate deadline;
+    private LocalDate deadline;
 
     public Deadline(String description, String deadline) {
         super(description);
