+++ src/main/java/duke/exceptions/DukeArrayOutOfBoundException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class DukeArrayOutOfBoundException extends DukeException{
