+++ src/main/java/duke/exceptions/DukeEmptyCommandException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class DukeEmptyCommandException extends DukeException{
