+++ src/main/java/duke/exceptions/DukeEmptyDescriptionException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class DukeEmptyDescriptionException extends DukeException{
