+++ src/main/java/duke/exceptions/DukeInvalidDescriptionException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class DukeInvalidDescriptionException extends DukeException {
