+++ src/main/java/duke/exceptions/DukeInvalidReadException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class DukeInvalidReadException extends DukeException{
