+++ src/main/java/duke/exceptions/DukeMissingIndexException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class DukeMissingIndexException extends DukeException {
