+++ src/main/java/duke/exceptions/DukeTooManyArgumentException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class DukeTooManyArgumentException extends DukeException{
