+++ src/main/java/duke/exceptions/DukeUnknownCommandException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class DukeUnknownCommandException extends DukeException{
